add upload page
I find an issue when I use iframe so we should find way to make the headbar and footer in all pages without error. Also we should restructuer the project files because it is complex to work with. And we should find the color root to be use by all of us.